Extract stateless AppState to separate module
