+++ src/client/src/app.ts
@@ -1,0 +1,26 @@
+import * as D from "./data";
+import * as T from "./tree";
+import * as Tutorial from "./tutorial";
+import {Communication} from "@thinktool/shared";
+
+export interface App {
+  state: D.State;
+  tutorialState: Tutorial.State;
+  changelogShown: boolean;
+  changelog: Communication.Changelog | "loading";
+  tree: T.Tree;
+}
+
+export function merge(app: App, values: {[K in keyof App]?: App[K]}): App {
+  const result = {...app};
+  for (const k in values) {
+    // I'm sure there's a smarter way to do this, but it doesn't really
+    // matter. Our function signature is fine.
+    (result[k as keyof App] as any) = values[k as keyof App]!;
+  }
+  return result;
+}
+
+export function jump(app: App, thing: string): App {
+  return merge(app, {tree: T.fromRoot(app.state, thing)});
+}
